+++ src/sys/mod.rs
@@ -1,5 +1,7 @@
 //! Functions with a common interface that rely on system calls.
 
+#![allow(unsafe_code)] // We're interfacing with system calls.
+
 #[cfg(feature = "local-offset")]
 mod local_offset_at;